langgraph-agent/requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-mcp-adapters>=0.1.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.0.0